Fixes a potential overflow when rescaling network evals.

+++ src/evaluate.cpp
@@ -25,6 +25,7 @@
 #include <memory>
 #include <sstream>
 
+#include "misc.h"
 #include "nnue/network.h"
 #include "nnue/nnue_misc.h"
 #include "position.h"
@@ -50,11 +51,11 @@ Value Eval::evaluate(const Eval::NNUE::Network&     network,
 
     // Blend optimism and eval with nnue complexity
     int nnueComplexity = std::abs(psqt - positional);
-    optimism += optimism * nnueComplexity / 465;
-    nnue -= nnue * nnueComplexity / 11743;
+    optimism += optimism * i64(nnueComplexity) / 465;
+    nnue -= nnue * i64(nnueComplexity) / 11743;
 
     int material = pos.major_material();
-    int v        = (nnue * (17380 + material) + optimism * (3061 + material)) / 20582;
+    int v        = (nnue * i64(17380 + material) + optimism * i64(3061 + material)) / 20582;
 
     // Damp down the evaluation linearly when shuffling
     v -= (v * pos.rule60_count()) / 253;
